fix some part of logout.

+++ src/main/java/app/store/service/dto/MediaDto.java
@@ -9,9 +9,9 @@ public class MediaDto {
     private AssetTypeDto type;
     private String name;
     private String url;
-    private Double capacity;
-    private Double width;
-    private Double height;
+    private Long capacity;
+    private Long width;
+    private Long height;
     private Double duration;
 
     public MediaDto() {
@@ -49,27 +49,27 @@ public class MediaDto {
         this.url = url;
     }
 
-    public Double getCapacity() {
+    public Long getCapacity() {
         return capacity;
     }
 
-    public void setCapacity(Double capacity) {
+    public void setCapacity(Long capacity) {
         this.capacity = capacity;
     }
 
-    public Double getWidth() {
+    public Long getWidth() {
         return width;
     }
 
-    public void setWidth(Double width) {
+    public void setWidth(Long width) {
         this.width = width;
     }
 
-    public Double getHeight() {
+    public Long getHeight() {
         return height;
     }
 
-    public void setHeight(Double height) {
+    public void setHeight(Long height) {
         this.height = height;
     }
